+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 2026-07-02
 
 - Lokalna baza ročišta sada se sprema šifrirano.
+- Restore backupa sada ima preview, test bez uvoza i sigurnu potvrdu prije obnove podataka.